Set up grid for collision detection.
I reorganized the current collision logic between the shapes and the
walls. It still isn't quite what it needs to be. I need to handle the
fact that the blocks have decimal positions that need rounded down.
First I need to make them all positive so that I don't have to do logic
on negative coords.