Set up rotation offsets.
This is done for shape T but not the others yet. It's worth noting that
this is a very bad implementation and if this project were likely to
last more than a couple weeks, I would put more thought into it. But
since I'm only interested in getting this working and finishing it, this
is good enough.